requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
future==1.0.0
geopandas==1.1.1
idna==3.11
kombu==5.5.4
lxml==6.0.2
Markdown==3.9
numpy==2.3.4
OWSLib==0.34.1
packaging==25.0
pandas==2.3.3
prompt_toolkit==3.0.52
pyogrio==0.11.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
serializers==0.2.4
shapely==2.1.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
